🐛 Improve cell renderer
